+++ p2p/client.go
@@ -1,11 +1,11 @@
 package p2p
 
 import (
+	"fmt"
 	"math"
 	"time"
 
 	"github.com/eoscanada/eos-go"
-	"github.com/pkg/errors"
 	"go.uber.org/zap"
 )
 
@@ -48,7 +48,7 @@ func (c *Client) read(peer *Peer, errChannel chan error) {
 	for {
 		packet, err := peer.Read()
 		if err != nil {
-			errChannel <- errors.Wrapf(err, "read message from %s", peer.Address)
+			errChannel <- fmt.Errorf("read message from %s: %w", peer.Address, err)
 			break
 		}
 
@@ -59,7 +59,7 @@ func (c *Client) read(peer *Peer, errChannel chan error) {
 
 		switch m := packet.P2PMessage.(type) {
 		case *eos.GoAwayMessage:
-			errChannel <- errors.Wrapf(err, "GoAwayMessage reason %s", m.Reason)
+			errChannel <- fmt.Errorf("GoAwayMessage reason %s: %w", m.Reason, err)
 
 		case *eos.HandshakeMessage:
 			if c.catchup == nil {
@@ -67,7 +67,7 @@ func (c *Client) read(peer *Peer, errChannel chan error) {
 				m.P2PAddress = peer.Name
 				err = peer.WriteP2PMessage(m)
 				if err != nil {
-					errChannel <- errors.Wrap(err, "HandshakeMessage")
+					errChannel <- fmt.Errorf("HandshakeMessage: %w", err)
 					break
 				}
 				zlog.Debug("Handshake resent", zap.String("other", m.P2PAddress))
@@ -77,7 +77,7 @@ func (c *Client) read(peer *Peer, errChannel chan error) {
 				c.catchup.originHeadBlock = m.HeadNum
 				err = c.catchup.sendSyncRequest(peer)
 				if err != nil {
-					errChannel <- errors.Wrap(err, "handshake: sending sync request")
+					errChannel <- fmt.Errorf("handshake: sending sync request: %w", err)
 				}
 				c.catchup.IsCatchingUp = true
 			}
@@ -88,7 +88,7 @@ func (c *Client) read(peer *Peer, errChannel chan error) {
 					c.catchup.originHeadBlock = pendingNum
 					err = c.catchup.sendSyncRequest(peer)
 					if err != nil {
-						errChannel <- errors.Wrap(err, "noticeMessage: sending sync request")
+						errChannel <- fmt.Errorf("noticeMessage: sending sync request: %w", err)
 					}
 				}
 			}
@@ -104,21 +104,21 @@ func (c *Client) read(peer *Peer, errChannel chan error) {
 						zlog.Debug("In sync with last handshake")
 						blockID, err := m.BlockID()
 						if err != nil {
-							errChannel <- errors.Wrap(err, "getting block id")
+							errChannel <- fmt.Errorf("getting block id: %w", err)
 						}
 						peer.handshakeInfo.HeadBlockNum = blockNum
 						peer.handshakeInfo.HeadBlockID = blockID
 						peer.handshakeInfo.HeadBlockTime = m.SignedBlockHeader.Timestamp.Time
 						err = peer.SendHandshake(peer.handshakeInfo)
 						if err != nil {
-							errChannel <- errors.Wrap(err, "send handshake")
+							errChannel <- fmt.Errorf("send handshake: %w", err)
 						}
 						zlog.Debug("Send new handshake",
 							zap.Object("handshakeInfo", peer.handshakeInfo))
 					} else {
 						err = c.catchup.sendSyncRequest(peer)
 						if err != nil {
-							errChannel <- errors.Wrap(err, "signed block: sending sync request")
+							errChannel <- fmt.Errorf("signed block: sending sync request: %w", err)
 						}
 					}
 				}
@@ -141,11 +141,11 @@ func (c *Client) Start() error {
 
 				err := triggerHandshake(c.peer)
 				if err != nil {
-					return errors.Wrap(err, "connect and start: trigger handshake")
+					return fmt.Errorf("connect and start: trigger handshake: %w", err)
 				}
 			}
 		case err := <-errorChannel:
-			return errors.Wrap(err, "start client")
+			return fmt.Errorf("start client: %w", err)
 		}
 	}
 }
@@ -173,7 +173,7 @@ func (c *Catchup) sendSyncRequest(peer *Peer) error {
 
 	err := peer.SendSyncRequest(c.requestedStartBlock, c.requestedEndBlock+1)
 	if err != nil {
-		return errors.Wrapf(err, "send sync request to %s", peer.Address)
+		return fmt.Errorf("send sync request to %s: %w", peer.Address, err)
 	}
 
 	return nil